Make suite result counters atomic; register cytoplast_manipulator
nbTests/nbFailed/nbAITests/nbAIFailed are incremented from every worker
thread, and the could-not-load reporting paths call handleResults
without the suite mutex - plain int increments could lose updates.
std::atomic<int> closes the hole (explicit casts at the fprintf summary;
atomics do not pass through varargs).

The historical "summary counted failures but no failure block printed"
anomaly no longer reproduces: across 19 recent full-suite logs the
summary count matches the printed blocks exactly, including the
nonzero case. It was a casualty of the since-fixed shared-state bugs
(thread-shared statics, the global RNG, the PTInstant UAF).

cytoplast_manipulator_i1085 was a finished, passing witness that was
never registered - enlisted (913 tests, registry 934). spell_rupture's
parked solo-hang also no longer reproduces; both witnesses pass solo in
seconds on the current engine.

+++ projects/mtg/src/GameStateDuel.cpp
@@ -925,7 +925,7 @@ void GameStateDuel::Update(float dt)
                     if (getenv("WAGIC_TESTSUITE"))
                     {
                         fprintf(stderr, "Test suite finished: %d tests (%d failed), %d AI tests (%d failed)\n",
-                                testSuite->nbTests, testSuite->nbFailed, testSuite->nbAITests, testSuite->nbAIFailed);
+                                (int)testSuite->nbTests, (int)testSuite->nbFailed, (int)testSuite->nbAITests, (int)testSuite->nbAIFailed);
                         exit((testSuite->nbFailed + testSuite->nbAIFailed) ? 1 : 0);
                     }
                     setGamePhase(DUEL_STATE_END);
